requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
langchain
pypdf2
fastapi
uvicorn[standard]
